+++ neuron/test/Neuron/Zettelkasten/Query/ParserSpec.hs
@@ -15,6 +15,7 @@ import qualified Network.URI.Encode as E
 import Neuron.Zettelkasten.Connection (Connection (..))
 import Neuron.Zettelkasten.ID (ZettelID (ZettelID))
 import Neuron.Zettelkasten.Query.Parser (parseQueryLink)
+import Neuron.Zettelkasten.Query.Theme (LinkView(LinkView_ShowDate), ZettelsView(ZettelsView))
 import Neuron.Zettelkasten.Zettel (ZettelQuery (..))
 import Relude
 import Test.Hspec
@@ -45,6 +46,15 @@ spec = do
     it "z:zettels?type=branch" $ do
       parseQueryLink (asURI "z:zettels?type=branch")
         `shouldBe` (Just $ Some $ ZettelQuery_ZettelsByTag [] Folgezettel def)
+    it "z:zettels?limit=10" $ do
+      parseQueryLink (asURI "z:zettels?limit=10")
+        `shouldBe` (Just $ Some $ ZettelQuery_ZettelsByTag [] OrdinaryConnection (ZettelsView def False $ Just 10))
+    it "z:zettels?timeline&limit=10" $ do
+      parseQueryLink (asURI "z:zettels?timeline&limit=10")
+        `shouldBe` (Just $ Some $ ZettelQuery_ZettelsByTag [] OrdinaryConnection (ZettelsView LinkView_ShowDate False $ Just 10))
+    it "z:zettels?limit=10&limit=20" $ do
+      parseQueryLink (asURI "z:zettels?limit=10&limit=20")
+        `shouldBe` (Just $ Some $ ZettelQuery_ZettelsByTag [] OrdinaryConnection (ZettelsView def False $ Just 10))
     it "z:tags" $ do
       parseQueryLink (asURI "z:tags")
         `shouldBe` (Just $ Some $ ZettelQuery_Tags [])